+++ src/snyk/common/api/apiСlient.ts
@@ -2,7 +2,7 @@ import axios, { AxiosError, AxiosInstance, AxiosRequestConfig, AxiosResponse } f
 import { IConfiguration } from '../configuration/configuration';
 import { configuration } from '../configuration/instance';
 import { ILog } from '../logger/interfaces';
-import { getAxiosProxyConfig } from '../proxy';
+import { getAxiosConfig } from '../proxy';
 import { IVSCodeWorkspace } from '../vscode/workspace';
 import { DEFAULT_API_HEADERS } from './headers';
 
@@ -11,7 +11,7 @@ export interface ISnykApiClient {
 }
 
 export class SnykApiClient implements ISnykApiClient {
-  private instance: AxiosInstance | null = null;
+  private instance: Promise<AxiosInstance> | null = null;
 
   constructor(
     private readonly configuration: IConfiguration,
@@ -19,16 +19,18 @@ export class SnykApiClient implements ISnykApiClient {
     private readonly logger: ILog,
   ) {}
 
-  private get http(): AxiosInstance {
+  private get http(): Promise<AxiosInstance> {
     return this.instance != null ? this.instance : this.initHttp();
   }
 
-  initHttp(): AxiosInstance {
-    const http = axios.create({
+  async initHttp(): Promise<AxiosInstance> {
+    const axiosRequestConfig: AxiosRequestConfig = {
       headers: DEFAULT_API_HEADERS,
       responseType: 'json',
-      ...getAxiosProxyConfig(this.workspace),
-    });
+      ...(await getAxiosConfig(this.workspace, this.configuration, this.logger)),
+    };
+
+    const http = axios.create(axiosRequestConfig);
 
     http.interceptors.response.use(
       response => response,
@@ -43,16 +45,16 @@ export class SnykApiClient implements ISnykApiClient {
       },
     );
 
-    this.instance = http;
+    this.instance = Promise.resolve(http);
     return http;
   }
 
   async get<T = unknown, R = AxiosResponse<T>>(url: string, config?: AxiosRequestConfig): Promise<R> {
-    return this.http.get<T, R>(url, await this.getRequestConfigWithAuth(config));
+    return (await this.http).get<T, R>(url, await this.getRequestConfigWithAuth(config));
   }
 
   async post<T = unknown, R = AxiosResponse<T>>(url: string, data: unknown, config?: AxiosRequestConfig): Promise<R> {
-    return this.http.post<T, R>(url, data, await this.getRequestConfigWithAuth(config));
+    return (await this.http).post<T, R>(url, data, await this.getRequestConfigWithAuth(config));
   }
 
   private async getRequestConfigWithAuth(config?: AxiosRequestConfig) {