+++ src/snyk/common/proxy.ts
@@ -1,26 +1,34 @@
 import { AxiosRequestConfig } from 'axios';
-import createHttpsProxyAgent, { HttpsProxyAgent, HttpsProxyAgentOptions } from 'https-proxy-agent';
+import fs from 'fs/promises';
+import { Agent, AgentOptions, globalAgent } from 'https';
+import { HttpsProxyAgent, HttpsProxyAgentOptions } from 'https-proxy-agent';
 import * as url from 'url';
 import { IVSCodeWorkspace } from './vscode/workspace';
+import { IConfiguration } from './configuration/configuration';
+import { ILog } from './logger/interfaces';
 
-export function getHttpsProxyAgent(
+export async function getHttpsProxyAgent(
   workspace: IVSCodeWorkspace,
+  configuration: IConfiguration,
+  logger: ILog,
   processEnv: NodeJS.ProcessEnv = process.env,
-): HttpsProxyAgent | undefined {
-  const proxyOptions = getProxyOptions(workspace, processEnv);
+): Promise<HttpsProxyAgent | undefined> {
+  const proxyOptions = await getProxyOptions(workspace, configuration, logger, processEnv);
   if (proxyOptions == undefined) return undefined;
 
   return new HttpsProxyAgent(proxyOptions);
 }
 
-export function getProxyOptions(
+export async function getProxyOptions(
   workspace: IVSCodeWorkspace,
+  configuration: IConfiguration,
+  logger: ILog,
   processEnv: NodeJS.ProcessEnv = process.env,
-): HttpsProxyAgentOptions | undefined {
+): Promise<HttpsProxyAgentOptions | undefined> {
   let proxy: string | undefined = getVsCodeProxy(workspace);
 
   const defaultOptions: HttpsProxyAgentOptions = {
-    rejectUnauthorized: workspace.getConfiguration<boolean>('http', 'proxyStrictSSL') ?? true,
+    ...(await getDefaultAgentOptions(configuration, logger)),
   };
 
   if (!proxy) {
@@ -44,6 +52,7 @@ export function getProxyOptions(
   if (proxyUrl.port && proxyUrl.port !== '') {
     port = parseInt(proxyUrl.port, 10);
   }
+
   return {
     host: proxyUrl.hostname,
     port: port,
@@ -57,17 +66,24 @@ export function getVsCodeProxy(workspace: IVSCodeWorkspace): string | undefined 
   return workspace.getConfiguration<string>('http', 'proxy');
 }
 
-export function getAxiosProxyConfig(workspace: IVSCodeWorkspace): AxiosRequestConfig {
+export async function getAxiosConfig(
+  workspace: IVSCodeWorkspace,
+  configuration: IConfiguration,
+  logger: ILog,
+): Promise<AxiosRequestConfig> {
+  // if proxying, we need to configure getHttpsProxyAgent, else configure getHttpsAgent
+  let agentOptions: HttpsProxyAgent | Agent | undefined = await getHttpsProxyAgent(workspace, configuration, logger);
+  if (!agentOptions) agentOptions = await getHttpsAgent(configuration, logger);
+
   return {
+    // proxy false as we're using https-proxy-agent library for proxying
     proxy: false,
-    httpAgent: getHttpsProxyAgent(workspace),
-    httpsAgent: getHttpsProxyAgent(workspace),
+    httpAgent: agentOptions,
+    httpsAgent: agentOptions,
   };
 }
 
-export function getProxyEnvVariable(
-  proxyOptions: createHttpsProxyAgent.HttpsProxyAgentOptions | undefined,
-): string | undefined {
+export function getProxyEnvVariable(proxyOptions: HttpsProxyAgentOptions | undefined): string | undefined {
   if (!proxyOptions) {
     return;
   }
@@ -76,4 +92,38 @@ export function getProxyEnvVariable(
 
   // noinspection HttpUrlsUsage
   return `${protocol}//${auth ? `${auth}@` : ''}${host}${port ? `:${port}` : ''}`;
+}
+
+async function getHttpsAgent(configuration: IConfiguration, logger: ILog): Promise<Agent> {
+  return new Agent({
+    ...(await getDefaultAgentOptions(configuration, logger)),
+  });
+}
+
+async function getDefaultAgentOptions(
+  configuration: IConfiguration,
+  logger: ILog,
+  processEnv: NodeJS.ProcessEnv = process.env,
+): Promise<AgentOptions | undefined> {
+  let defaultOptions: AgentOptions | undefined;
+
+  const sslCheck = !configuration.getInsecure();
+  defaultOptions = { rejectUnauthorized: sslCheck };
+  if (!sslCheck) {
+    globalAgent.options.rejectUnauthorized = false;
+  } else {
+    // use custom certs if provided
+    if (processEnv.NODE_EXTRA_CA_CERTS) {
+      try {
+        await fs.access(processEnv.NODE_EXTRA_CA_CERTS);
+        const certs = await fs.readFile(processEnv.NODE_EXTRA_CA_CERTS);
+        defaultOptions = { ca: [certs] };
+        globalAgent.options.ca = [certs];
+      } catch (error) {
+        logger.error(`Failed to read NODE_EXTRA_CA_CERTS file: ${error}`);
+      }
+    }
+  }
+
+  return defaultOptions;
 }