+++ src/snyk/snykOss/api/npmTestApi.ts
@@ -2,24 +2,24 @@ import axios, { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
 import { DEFAULT_API_HEADERS } from '../../common/api/headers';
 import { configuration } from '../../common/configuration/instance';
 import { ILog } from '../../common/logger/interfaces';
-import { getAxiosProxyConfig } from '../../common/proxy';
+import { getAxiosConfig } from '../../common/proxy';
 import { IVSCodeWorkspace } from '../../common/vscode/workspace';
 
 export class NpmTestApi {
-  private instance: AxiosInstance | null = null;
+  private instance: Promise<AxiosInstance> | null = null;
 
   constructor(private readonly logger: ILog, private readonly workspace: IVSCodeWorkspace) {}
 
-  private get http(): AxiosInstance {
+  private get http(): Promise<AxiosInstance> {
     return this.instance != null ? this.instance : this.initHttp();
   }
 
-  initHttp(): AxiosInstance {
+  async initHttp(): Promise<AxiosInstance> {
     const http = axios.create({
       headers: DEFAULT_API_HEADERS,
       responseType: 'json',
       baseURL: configuration.authHost + '/test',
-      ...getAxiosProxyConfig(this.workspace),
+      ...(await getAxiosConfig(this.workspace, configuration, this.logger)),
     });
 
     http.interceptors.response.use(
@@ -30,11 +30,11 @@ export class NpmTestApi {
       },
     );
 
-    this.instance = http;
+    this.instance = Promise.resolve(http);
     return http;
   }
 
-  get<T = unknown, R = AxiosResponse<T>>(url: string, config?: AxiosRequestConfig): Promise<R> {
-    return this.http.get<T, R>(url, config);
+  async get<T = unknown, R = AxiosResponse<T>>(url: string, config?: AxiosRequestConfig): Promise<R> {
+    return (await this.http).get<T, R>(url, config);
   }
 }