+++ src/test/unit/common/proxy.test.ts
@@ -3,8 +3,10 @@
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 import assert from 'assert';
 import sinon from 'sinon';
-import { getHttpsProxyAgent, getProxyEnvVariable, getProxyOptions } from '../../../snyk/common/proxy';
+import { getAxiosConfig, getHttpsProxyAgent, getProxyEnvVariable, getProxyOptions } from '../../../snyk/common/proxy';
 import { IVSCodeWorkspace } from '../../../snyk/common/vscode/workspace';
+import { IConfiguration } from '../../../snyk/common/configuration/configuration';
+import { ILog } from '../../../snyk/common/logger/interfaces';
 
 suite('Proxy', () => {
   // Proxy settings
@@ -14,25 +16,77 @@ suite('Proxy', () => {
   const protocol = 'https:';
   const proxy = `${protocol}//${auth}@${host}:${port}`;
   const proxyStrictSSL = true;
+  const error = sinon.stub();
+  const logger = { error } as unknown as ILog;
 
   teardown(() => {
     sinon.restore();
   });
 
-  test('No proxy set', () => {
+  test('No proxy configured by user (default case)', async () => {
     const getConfiguration = sinon.stub();
+    const getInsecure = sinon.stub();
+    getInsecure.returns(!proxyStrictSSL);
+
+    const configuration = {
+      getInsecure,
+    } as unknown as IConfiguration;
+
     getConfiguration.withArgs('http', 'proxy').returns(undefined);
 
     const workspace = {
       getConfiguration,
     } as unknown as IVSCodeWorkspace;
 
-    const agent = getHttpsProxyAgent(workspace);
+    const configOptions = await getAxiosConfig(workspace, configuration, logger);
 
-    assert.deepStrictEqual(agent, undefined);
+    // should still set rejectUnauthorized flag
+    // @ts-ignore: cannot test options otherwise
+    assert.deepStrictEqual(configOptions.httpAgent?.options.rejectUnauthorized, proxyStrictSSL);
+    assert.deepStrictEqual(configOptions.httpsAgent?.options.rejectUnauthorized, proxyStrictSSL);
   });
 
-  test('Proxy is set in VS Code settings', () => {
+  suite('.getAxiosConfig()', () => {
+    suite('when proxyStrictSsl checkbox is checked', () => {
+      const getConfiguration = sinon.stub();
+      const workspace = {
+        getConfiguration,
+      } as unknown as IVSCodeWorkspace;
+
+      const getInsecure = sinon.stub();
+      getInsecure.returns(false);
+
+      const configuration = {
+        getInsecure,
+      } as unknown as IConfiguration;
+
+      test('should return rejectUnauthorized true', async () => {
+        const config = await getAxiosConfig(workspace, configuration, logger);
+        assert.deepStrictEqual(config.httpAgent?.options.rejectUnauthorized, true);
+      });
+    });
+
+    suite('when proxyStrictSsl checkbox is not checked', () => {
+      const getConfiguration = sinon.stub();
+      const workspace = {
+        getConfiguration,
+      } as unknown as IVSCodeWorkspace;
+
+      const getInsecure = sinon.stub();
+      getInsecure.returns(true);
+
+      const configuration = {
+        getInsecure,
+      } as unknown as IConfiguration;
+
+      test('should return rejectUnauthorized false', async () => {
+        const config = await getAxiosConfig(workspace, configuration, logger);
+        assert.deepStrictEqual(config.httpsAgent?.options.rejectUnauthorized, false);
+      });
+    });
+  });
+
+  test('Proxy is configured in VS Code settings', async () => {
     const getConfiguration = sinon.stub();
     getConfiguration.withArgs('http', 'proxy').returns(proxy);
     getConfiguration.withArgs('http', 'proxyStrictSSL').returns(proxyStrictSSL);
@@ -41,7 +95,14 @@ suite('Proxy', () => {
       getConfiguration,
     } as unknown as IVSCodeWorkspace;
 
-    const agent = getHttpsProxyAgent(workspace);
+    const getInsecure = sinon.stub();
+    getInsecure.returns(!proxyStrictSSL);
+
+    const configuration = {
+      getInsecure,
+    } as unknown as IConfiguration;
+
+    const agent = await getHttpsProxyAgent(workspace, configuration, logger);
 
     // @ts-ignore: cannot test options otherwise
     assert.deepStrictEqual(agent?.proxy.host, host);
@@ -53,7 +114,7 @@ suite('Proxy', () => {
     assert.deepStrictEqual(agent?.proxy.rejectUnauthorized, proxyStrictSSL);
   });
 
-  test('Proxy is set in environment', () => {
+  test('Proxy is configured in environment', async () => {
     const getConfiguration = sinon.stub();
     getConfiguration.withArgs('http', 'proxy').returns(undefined);
     getConfiguration.withArgs('http', 'proxyStrictSSL').returns(proxyStrictSSL);
@@ -62,7 +123,13 @@ suite('Proxy', () => {
       getConfiguration,
     } as unknown as IVSCodeWorkspace;
 
-    const agent = getHttpsProxyAgent(workspace, {
+    const getInsecure = sinon.stub();
+    getInsecure.returns(!proxyStrictSSL);
+    const configuration = {
+      getInsecure,
+    } as unknown as IConfiguration;
+
+    const agent = await getHttpsProxyAgent(workspace, configuration, logger, {
       https_proxy: proxy,
       http_proxy: proxy,
     });
@@ -77,7 +144,7 @@ suite('Proxy', () => {
     assert.deepStrictEqual(agent?.proxy.rejectUnauthorized, proxyStrictSSL);
   });
 
-  test('getProxyEnvVariable should return the https proxy as env var', () => {
+  test('getProxyEnvVariable should return the https proxy as env var', async () => {
     const getConfiguration = sinon.stub();
     getConfiguration.withArgs('http', 'proxy').returns(proxy);
     getConfiguration.withArgs('http', 'proxyStrictSSL').returns(proxyStrictSSL);
@@ -86,7 +153,14 @@ suite('Proxy', () => {
       getConfiguration,
     } as unknown as IVSCodeWorkspace;
 
-    const envVariable = getProxyEnvVariable(getProxyOptions(workspace));
+    const getInsecure = sinon.stub();
+    getInsecure.returns(!proxyStrictSSL);
+
+    const configuration = {
+      getInsecure,
+    } as unknown as IConfiguration;
+
+    const envVariable = getProxyEnvVariable(await getProxyOptions(workspace, configuration, logger));
 
     // noinspection HttpUrlsUsage
     assert.deepStrictEqual(envVariable, `${protocol}//${auth}@${host}:${port}`);